+++ cggmp21/src/key_refresh/aux_only.rs
@@ -26,6 +26,12 @@ use crate::{
 
 use super::{Bug, KeyRefreshError, PregeneratedPrimes, ProtocolAborted};
 
+macro_rules! prefixed {
+    ($name:tt) => {
+        concat!("dfns.cggmp21.aux_gen.", $name)
+    };
+}
+
 /// Message of key refresh protocol
 #[derive(ProtocolMessage, Clone, Serialize, Deserialize)]
 #[serde(bound = "")]
@@ -44,7 +50,7 @@ pub enum Msg<D: Digest, L: SecurityLevel> {
 
 /// Message from round 1
 #[derive(Clone, Serialize, Deserialize, udigest::Digestable)]
-#[udigest(tag = "dfns.cggmp21.aux_gen.round1")]
+#[udigest(tag = prefixed!("round1"))]
 #[udigest(bound = "")]
 #[serde(bound = "")]
 pub struct MsgRound1<D: Digest> {
@@ -54,7 +60,7 @@ pub struct MsgRound1<D: Digest> {
 }
 /// Message from round 2
 #[derive(Clone, Serialize, Deserialize, udigest::Digestable)]
-#[udigest(tag = "dfns.cggmp21.aux_gen.round2")]
+#[udigest(tag = prefixed!("round2"))]
 #[udigest(bound = "")]
 #[serde(bound = "")]
 pub struct MsgRound2<L: SecurityLevel> {
@@ -98,20 +104,53 @@ pub struct MsgRound3 {
 #[serde(bound = "")]
 pub struct MsgReliabilityCheck<D: Digest>(pub digest::Output<D>);
 
-#[derive(udigest::Digestable)]
-#[udigest(tag = "dfns.cggmp21.aux_gen.tag")]
-enum Tag<'a> {
-    /// Tag that includes the prover index
-    Indexed {
-        party_index: u16,
+mod unambiguous {
+    use digest::Digest;
+
+    use crate::{ExecutionId, SecurityLevel};
+
+    #[derive(udigest::Digestable)]
+    #[udigest(tag = prefixed!("proof_prm"))]
+    pub struct ProofPrm<'a> {
+        pub sid: ExecutionId<'a>,
+        pub prover: u16,
+    }
+
+    #[derive(udigest::Digestable)]
+    #[udigest(tag = prefixed!("proof_mod"))]
+    pub struct ProofMod<'a> {
+        pub sid: ExecutionId<'a>,
         #[udigest(as_bytes)]
-        sid: &'a [u8],
-    },
-    /// Tag w/o party index
-    Unindexed {
+        pub rho: &'a [u8],
+        pub prover: u16,
+    }
+
+    #[derive(udigest::Digestable)]
+    #[udigest(tag = prefixed!("proof_fac"))]
+    #[udigest(bound = "")]
+    pub struct ProofFac<'a> {
+        pub sid: ExecutionId<'a>,
         #[udigest(as_bytes)]
-        sid: &'a [u8],
-    },
+        pub rho: &'a [u8],
+        pub prover: u16,
+    }
+
+    #[derive(udigest::Digestable)]
+    #[udigest(tag = prefixed!("hash_commitment"))]
+    #[udigest(bound = "")]
+    pub struct HashCom<'a, L: SecurityLevel> {
+        pub sid: ExecutionId<'a>,
+        pub prover: u16,
+        pub decommitment: &'a super::MsgRound2<L>,
+    }
+
+    #[derive(udigest::Digestable)]
+    #[udigest(tag = prefixed!("echo_round"))]
+    #[udigest(bound = "")]
+    pub struct Echo<'a, D: Digest> {
+        pub sid: ExecutionId<'a>,
+        pub commitment: &'a super::MsgRound1<D>,
+    }
 }
 
 pub async fn run_aux_gen<R, M, L, D>(
@@ -119,7 +158,7 @@ pub async fn run_aux_gen<R, M, L, D>(
     n: u16,
     mut rng: &mut R,
     party: M,
-    execution_id: ExecutionId<'_>,
+    sid: ExecutionId<'_>,
     pregenerated: PregeneratedPrimes<L>,
     mut tracer: Option<&mut dyn Tracer>,
     reliable_broadcast_enforced: bool,
@@ -147,17 +186,6 @@ where
     let round3 = rounds.add_round(RoundInput::<MsgRound3>::p2p(i, n));
     let mut rounds = rounds.listen(incomings);
 
-    tracer.stage("Precompute execution id and shared state");
-    let sid = execution_id.as_bytes();
-    let tag = |j| {
-        udigest::Tag::<D>::new_structured(Tag::Indexed {
-            party_index: j,
-            sid,
-        })
-    };
-    let tag_i = tag(i);
-    let parties_shared_state = D::new_with_prefix(D::digest(sid));
-
     // Round 1
     tracer.round_begins();
 
@@ -176,8 +204,8 @@ where
     let s = t.pow_mod_ref(&lambda, &N).ok_or(Bug::PowMod)?.into();
 
     tracer.stage("Prove Πprm (ψˆ_i)");
-    let hat_psi = π_prm::prove(
-        parties_shared_state.clone().chain_update(i.to_be_bytes()),
+    let hat_psi = π_prm::prove::<{ crate::security_level::M }, D>(
+        &unambiguous::ProofPrm { sid, prover: i },
         &mut rng,
         π_prm::Data {
             N: &N,
@@ -208,7 +236,11 @@ where
             nonce
         },
     };
-    let hash_commit = tag_i.clone().digest(&decommitment);
+    let hash_commit = udigest::hash::<D>(&unambiguous::HashCom {
+        sid,
+        prover: i,
+        decommitment: &decommitment,
+    });
 
     tracer.send_msg();
     let commitment = MsgRound1 {
@@ -233,8 +265,11 @@ where
     // Optional reliability check
     if reliable_broadcast_enforced {
         tracer.stage("Hash received msgs (reliability check)");
-        let h_i = udigest::Tag::<D>::new_structured(Tag::Unindexed { sid })
-            .digest_iter(commitments.iter_including_me(&commitment));
+        let h_i = udigest::hash_iter::<D>(
+            commitments
+                .iter_including_me(&commitment)
+                .map(|commitment| unambiguous::Echo { sid, commitment }),
+        );
 
         tracer.send_msg();
         outgoings
@@ -285,7 +320,12 @@ where
     // validate decommitments
     tracer.stage("Validate round 1 decommitments");
     let blame = collect_blame(&decommitments, &commitments, |j, decomm, comm| {
-        tag(j).digest(decomm) != comm.commitment
+        let com_expected = udigest::hash::<D>(&unambiguous::HashCom {
+            sid,
+            prover: j,
+            decommitment: decomm,
+        });
+        com_expected != comm.commitment
     });
     if !blame.is_empty() {
         return Err(ProtocolAborted::invalid_decommitment(blame).into());
@@ -301,8 +341,8 @@ where
                 s: &d.s,
                 t: &d.t,
             };
-            π_prm::verify(
-                parties_shared_state.clone().chain_update(j.to_be_bytes()),
+            π_prm::verify::<{ crate::security_level::M }, D>(
+                &unambiguous::ProofPrm { sid, prover: j },
                 data,
                 &d.params_proof,
             )
@@ -321,13 +361,13 @@ where
         .fold(rho_bytes, utils::xor_array);
 
     // common data for messages
-    let my_shared_state = parties_shared_state
-        .clone()
-        .chain_update(i.to_be_bytes())
-        .chain_update(&rho_bytes);
     tracer.stage("Compute П_mod (ψ_i)");
-    let psi = π_mod::non_interactive::prove(
-        my_shared_state.clone(),
+    let psi = π_mod::non_interactive::prove::<{ crate::security_level::M }, D>(
+        &unambiguous::ProofMod {
+            sid,
+            rho: rho_bytes.as_ref(),
+            prover: i,
+        },
         &π_mod::Data { n: N.clone() },
         &π_mod::PrivateData {
             p: p.clone(),
@@ -349,8 +389,12 @@ where
         tracer.send_msg();
 
         tracer.stage("Compute П_fac (ф_i^j)");
-        let phi = π_fac::prove(
-            my_shared_state.clone(),
+        let phi = π_fac::prove::<D>(
+            &unambiguous::ProofFac {
+                sid,
+                rho: rho_bytes.as_ref(),
+                prover: i,
+            },
             &π_fac::Aux {
                 s: d.s.clone(),
                 t: d.t.clone(),
@@ -400,11 +444,12 @@ where
                 n: decommitment.N.clone(),
             };
             let (comm, proof) = &proof_msg.mod_proof;
-            π_mod::non_interactive::verify(
-                parties_shared_state
-                    .clone()
-                    .chain_update(j.to_be_bytes())
-                    .chain_update(&rho_bytes),
+            π_mod::non_interactive::verify::<{ crate::security_level::M }, D>(
+                &unambiguous::ProofMod {
+                    sid,
+                    rho: rho_bytes.as_ref(),
+                    prover: j,
+                },
                 &data,
                 comm,
                 proof,
@@ -436,11 +481,12 @@ where
         &decommitments,
         &shares_msg_b,
         |j, decommitment, proof_msg| {
-            π_fac::verify(
-                parties_shared_state
-                    .clone()
-                    .chain_update(j.to_be_bytes())
-                    .chain_update(&rho_bytes),
+            π_fac::verify::<D>(
+                &unambiguous::ProofFac {
+                    sid,
+                    rho: rho_bytes.as_ref(),
+                    prover: j,
+                },
                 &phi_common_aux,
                 π_fac::Data {
                     n: &decommitment.N,